refactor: move moving things around
