+++ libs/wallet-react/src/components/connect-dialog/index.ts
@@ -2,5 +2,5 @@ export { ConnectDialogWithRiskAck, ConnectDialog } from './connect-dialog';
 export { ConnectionOptions } from './connection-options';
 export { ConnectionStatus } from './connection-status';
 export { ConnectorIcon } from './connection-options/connector-icon';
-export { ConnectionOptionDefault } from './connection-options/connection-option-default';
-export { ConnectionOptionRecord } from './connection-options/custom-connection-options';
+export { ConnectionOptionDefault } from './connection-options';
+export { ConnectionOptionRecord } from './connection-options';
